refactor(plugin-zod): share printer helpers between printerZod and printerZodMini
Extract the traversal and formatting logic duplicated between the two zod
printers into printers/shared.ts, parameterized by a ZodDialect that carries
the emissions where the chainable and zod/mini APIs differ. Each printer keeps
only its dialect and the genuinely divergent node handlers (string, number,
integer, bigint, date, datetime, time). Generated output is unchanged.

Closes #674
